refactor: LottoFactory 클래스 분리
